Separate `SearchesAdapter` and `MoreSearchAdapter` for now.
