+++ graylog2-server/src/main/java/org/graylog/events/search/MoreSearchAdapter.java
@@ -1,0 +1,19 @@
+package org.graylog.events.search;
+
+import org.graylog.events.processor.EventProcessorException;
+import org.graylog2.indexer.results.ResultMessage;
+import org.graylog2.indexer.searches.Sorting;
+import org.graylog2.plugin.indexer.searches.timeranges.TimeRange;
+
+import java.util.List;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+public interface MoreSearchAdapter {
+    MoreSearch.Result eventSearch(String queryString, TimeRange timerange, Set<String> affectedIndices, Sorting sorting, int page, int perPage, Set<String> eventStreams, String filterString, Set<String> forbiddenSourceStreams);
+
+    interface ScrollEventsCallback {
+        void accept(List<ResultMessage> results, AtomicBoolean requestContinue) throws EventProcessorException;
+    }
+    void scrollEvents(String queryString, TimeRange timeRange, Set<String> affectedIndices, Set<String> streams, String scrollTime, int batchSize, ScrollEventsCallback resultCallback) throws EventProcessorException;
+}
